requirements_freeze.txt
flask
PyCrypto
py2neo
nose
mock
PyHamcrest